requirements.txt
gradio==3.50.2
mysql-connector-python==8.0.33
requests==2.31.0
Pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0